Refactoring and add workflow file
