+++ social-scout-twitter/src/main/java/echoic/socialscouttwitter/core/MusicEntityGeneratorImpl.java
@@ -1,5 +1,6 @@
 package echoic.socialscouttwitter.core;
 
+import echoic.socialscouttwitter.core.interfaces.MusicEntityGenerator;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
