--- social-scout-twitter/src/main/java/echoic/socialscouttwitter/core/QueryTokenizer.java
@@ -1,6 +1,0 @@
-package echoic.socialscouttwitter.core;
-
-public interface QueryTokenizer
-{
-    String getSearchString(String post) throws TokenizationException;
-}
