+++ social-scout-twitter/src/main/java/echoic/socialscouttwitter/core/QueryTokenizerImp.java
@@ -1,13 +1,7 @@
 package echoic.socialscouttwitter.core;
 
+import echoic.socialscouttwitter.core.interfaces.QueryTokenizer;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.StringUtils;
-import twitter4j.Status;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 @Slf4j
 public class QueryTokenizerImp implements QueryTokenizer {
