+++ social-scout-twitter/src/main/java/echoic/socialscouttwitter/core/interfaces/QueryTokenizer.java
@@ -1,0 +1,8 @@
+package echoic.socialscouttwitter.core.interfaces;
+
+import echoic.socialscouttwitter.core.TokenizationException;
+
+public interface QueryTokenizer
+{
+    String getSearchString(String post) throws TokenizationException;
+}
