+++ social-scout-twitter/src/main/java/echoic/socialscouttwitter/post/interfaces/Tweeter.java
@@ -1,4 +1,4 @@
-package echoic.socialscouttwitter.post;
+package echoic.socialscouttwitter.post.interfaces;
 
 import echoic.socialscouttwitter.core.MusicEntity;
 import twitter4j.Status;
