+++ social-scout-twitter/src/test/java/echoic/socialscouttwitter/core/MusicEntityGeneratorTest.java
@@ -1,5 +1,6 @@
 package echoic.socialscouttwitter.core;
 
+import echoic.socialscouttwitter.core.interfaces.MusicEntityGenerator;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
@@ -11,11 +12,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestTemplate;
 
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.Mockito.when;
