+++ social-scout-twitter/src/test/java/echoic/socialscouttwitter/core/PostHandlerTest.java
@@ -1,23 +1,18 @@
 package echoic.socialscouttwitter.core;
 
-import echoic.socialscouttwitter.post.Tweeter;
+import echoic.socialscouttwitter.core.interfaces.MusicEntityGenerator;
+import echoic.socialscouttwitter.core.interfaces.PostHandler;
+import echoic.socialscouttwitter.core.interfaces.QueryTokenizer;
+import echoic.socialscouttwitter.post.interfaces.Tweeter;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import twitter4j.Status;
-import twitter4j.Twitter;
 
-import java.util.Map;
-
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
