+++ social-scout-twitter/src/test/java/echoic/socialscouttwitter/core/QueryTokenizerTest.java
@@ -1,13 +1,11 @@
 package echoic.socialscouttwitter.core;
 
+import echoic.socialscouttwitter.core.interfaces.QueryTokenizer;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 class QueryTokenizerTest {
